Add RaiseCeiling & several control dvars

+++ src/Components/Modules/IMapEnts.cpp
@@ -23,10 +23,14 @@ namespace Components
 		mapEnts.DeleteOldSchoolPickups();
 		mapEnts.AddRemovedSModels();
 
-		if (mapEnts.ConvertTurrets())
-		{
-			// The map features turrets! We need to write a file somewhere to inform the converter about it so that the proper iw4 minigun source will be included
-			Utils::WriteFile(Utils::VA("%s/HAS_MINIGUN", AssetHandler::GetExportPath().data()), "\0");
+
+		static const auto replaceTurretsWithMiniguns = Game::Dvar_FindVar("iw3x_replace_turrets_with_miniguns");
+		if (replaceTurretsWithMiniguns && replaceTurretsWithMiniguns->current.string == "1"s) {
+			if (mapEnts.ConvertTurrets())
+			{
+				// The map features turrets! We need to write a file somewhere to inform the converter about it so that the proper iw4 minigun source will be included
+				Utils::WriteFile(Utils::VA("%s/HAS_MINIGUN", AssetHandler::GetExportPath().data()), "\0");
+			}
 		}
 
 		bool hasVehicles = mapEnts.ConvertVehicles();
@@ -38,8 +42,8 @@ namespace Components
 
 		iw4Ents->name = ents->name;
 		iw4Ents->entityString = LocalAllocator.DuplicateString(entString);
-		iw4Ents->numEntityChars = entString.size() +1;
-		iw4Ents->stages = nullptr; 
+		iw4Ents->numEntityChars = entString.size() + 1;
+		iw4Ents->stages = nullptr;
 
 
 		return iw4Ents;
@@ -50,17 +54,17 @@ namespace Components
 		Command::Add("dumpMapEnts", [](const Command::Params& params)
 			{
 				if (params.Length() < 2) return;
-				 
-				 auto converted = IMapEnts::Convert(Game::DB_FindXAssetHeader(Game::IW3::XAssetType::ASSET_TYPE_CLIPMAP_PVS, params[1]).clipMap->mapEnts);
-				 MapDumper::GetApi()->write(Game::IW4::ASSET_TYPE_MAP_ENTS, converted);
-				 
-				 std::string entString(converted->entityString, converted->numEntityChars - 1);
-				 Utils::Entities mapEnts(entString);
-				 const auto models = mapEnts.GetModels(false);
-				 for (const std::string& model : models)
-				 {
-					 Command::Execute(Utils::VA("dumpXModel %s", model.data()), true);
-				 }
+
+				auto converted = IMapEnts::Convert(Game::DB_FindXAssetHeader(Game::IW3::XAssetType::ASSET_TYPE_CLIPMAP_PVS, params[1]).clipMap->mapEnts);
+				MapDumper::GetApi()->write(Game::IW4::ASSET_TYPE_MAP_ENTS, converted);
+
+				std::string entString(converted->entityString, converted->numEntityChars - 1);
+				Utils::Entities mapEnts(entString);
+				const auto models = mapEnts.GetModels(false);
+				for (const std::string& model : models)
+				{
+					Command::Execute(Utils::VA("dumpXModel %s", model.data()), true);
+				}
 			});
 	}
